Fix integer overflow in mpp_table_size view for large tables
The size_mb expression computed `c.relpages * 32/1024` in int4
arithmetic. Since pg_class.relpages is int4 and the literal 32 is
int4, `relpages * 32` overflows int4 once relpages exceeds ~67M
pages (a heap table of roughly 2TB at the 32KB block size),
raising "ERROR: integer out of range". This aborted the Database
Size check (and any check reading this view) on clusters holding
large tables.

Cast relpages to bigint before the multiplication and make the
whole size_mb column bigint so both CASE branches share one type.
Verified on SynxDB4 4.5.2: the old expression errors at
relpages=70000000 while the new one returns 2187500 MB, and the
db/schema/table size queries all run clean.

+++ synxdb_health_check.py
@@ -183,8 +183,8 @@ SELECT distinct dfhostname, dfdevice, (dfspace/1024/1024)::decimal(18,2) as "spa
 create_mpp_table_size_view_sql = '''
 create or replace view public.mpp_table_size
 as select c.oid,n.nspname as schemaname,c.relname as tablename,
-(case when c.relpages > 0 then c.relpages * 32/1024
-      else (pg_relation_size(c.oid)/1024/1024)::int end) as size_mb
+(case when c.relpages > 0 then c.relpages::bigint * 32/1024
+      else (pg_relation_size(c.oid)/1024/1024) end)::bigint as size_mb
 from pg_class c join pg_namespace n on c.relnamespace=n.oid
 where c.relkind='r'
 '''
